feature: Actualizar header para visualizar y editar foto de perfil tanto para turista como para proveedor (#23)

+++ front-I-Wellness/src/app/features/users/proveedor/header-proveedor/header-proveedor.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
+import { AuthService } from '../../../../core/services/auth/auth.service';
 
 @Component({
   selector: 'app-header-proveedor',
@@ -8,9 +9,31 @@ import { Router } from '@angular/router';
   styleUrl: './header-proveedor.component.css'
 })
 export class HeaderProveedorComponent {
-    constructor(private router: Router) {}
-    navigateTo(path: string) {
+  usuario: any;
+  constructor(private router: Router, private authService: AuthService) {}
+
+  ngOnInit(): void {
+    this.authService.usuarioHome().subscribe({
+      next: (data: string) => {
+        this.usuario = data;
+        this.usuario = JSON.parse(data);
+        console.log(this.usuario)
+      },
+      error: (err: any) => {
+        console.error('Error al obtener el usuario:', err);
+      }
+    });
+  }
+
+  navigateTo(path: string, param?: any) {
+    // Si hay un parámetro y no es undefined, navegar con parámetro
+    if (param !== undefined) {
+      this.router.navigate([path, param]);
+    } else {
+      // Si no hay parámetro o es undefined, navegar solo con la ruta
       localStorage.clear();
       this.router.navigate([path]);
     }
+  }
+   
 }
